backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.3
pydantic-settings==2.6.1
sqlalchemy==2.0.36
pymysql==1.1.1
redis==5.2.1
python-multipart==0.0.17
httpx==0.27.2
requests==2.32.3
PyYAML==6.0.2
python-dotenv==1.0.1
qcloud_cos==5.0.5
edge-tts==6.1.12
openai==1.58.1
vosk==0.3.45
PyJWT==2.10.1
numpy==2.1.3
onnxruntime==1.20.1
opuslib==3.0.1
dashscope==1.20.6
tencentcloud-sdk-python==3.0.1310
alibabacloud-dysmsapi20170525==2.0.24
alibabacloud-tea-openapi==0.3.9
websocket-client==1.8.0
